Get conditional responses from a GP model

This will obtain responses to each predictor variable across its range of observed values, conditional on all other predictor variables being set to 0 (their mean value, if data are scaled properly). If automatic scaling is used, predictors and responses are backtransformed to their original scale. If the model includes multiple populations, responses are obtained for each population.

getconditionals(fit, xrange = "default", extrap = 0.01, nvals = 25, plot = T)

Arguments

fit

Output from fitGP.

xrange

The range of scaled predictor values over which responses should be evaluated. Options are "local" (the range observed within each population), or "global" (the range across all populations). Irrelevant if there is only one population. The default ("default") will be "local" if scaling="local" was used, and "global" if scaling="global" or scaling="none" was used. This option can useful for either preventing or allowing for extrapolation beyond the range of the data in a given population.

extrap

Percentage to extrapolate beyond the predictor range for each population (i.e. predictor values will range from (1-extrap)*min(X) to (1+extrap)*max(X) Defaults to 0.01.

nvals

The number of values to evaluate across the predictor range for each population. Defaults to 25, which should be fine for most applications.

plot

Produce a plot, or not (logical, defaults to TRUE).

Value

Returns (invisibly) a data frame containing the predictor values and conditional responses (means and standard deviations).