Generate delay vectors

Create a lag matrix for a time series (or suite of time series) using given E and tau values. Like in fitGP, either a data frame and column names or vectors/matrices can be provided. Use forecast=TRUE to generate a forecast matrix. Use vtimestep=TRUE to use the variable timestep method.

makelags(
  data = NULL,
  y,
  pop = NULL,
  E,
  tau,
  yname = NULL,
  forecast = FALSE,
  vtimestep = FALSE,
  x = NULL,
  time = NULL,
  augment = FALSE,
  Tdiff_max = NULL,
  Tdiff_fore = NULL,
  nreps = NULL,
  append = FALSE
)

Arguments

data

A data frame, or matrix with named columns.

y

A vector containing a time series (usually the response variable). Alternatively, a matrix or data frame where each column is a time series (usually the response variable and covariates). In the case of multiple time series, lags will be generated for each variable. If data is supplied, the column names or indices.

pop

A vector of populations (optional, if not supplied defaults to 1 population). If data is supplied, the column name or index.

E

Embedding dimension. Required.

tau

Time delay. Required.

yname

Optional, name of the variable if y is input as a vector, or variables if y is a matrix with unnamed columns. Otherwise not used. Defaults to "var" if NULL.

forecast

Produce a forecast matrix instead of the standard training/test matrix.

vtimestep

Use variable timestep method.

x

If using vtimestep=TRUE, put the response variable under use y, and covariates under x (see details). If vtimestep=FALSE, will be appended to y.

time

Used to generate forecast time if forecast=TRUE and to calculate Tdiff if vtimestep=TRUE. If not supplied and vtimestep=TRUE, a numeric index is used.

augment

If vtimestep=TRUE, produce augmentation lag matrix.

Tdiff_max

If vtimestep=TRUE, the maximum Tdiff value to allow in production of the lag matrix or augmentation lag matrix.

Tdiff_fore

If vtimestep=TRUE and forecast=TRUE, vector of time units to forecast beyond the last timestep. Defaults to the minimum time difference between consecutive timepoints, multipled by tau.

nreps

If augment=TRUE, the max number of delay vectors for each Tdiff value.

append

Return data with the lags appended (defaults to FALSE). Only relevant if forecast=FALSE and augment=FALSE.

Value

A matrix with named columns, the appended number indicating the time lag. If y has named columns, named columns in the lag matrix will match. If generating forecasts, the output matrix will include a column for population if there is more than one, and will include a time column if time is supplied. If using the variable timestep method, the output matrix will include lags of Tdiff (time difference).

Details

When using the standard (fixed timestep) method, the response variable and covariates can all be input under y, and lags will be generated for all variables.

When using the variable timestep method, it is necessary to differentiate between the response variable and covariates, as they are handled differently. The response variable should go under y and the covariates should go under x. The output matrix will include lags of Tdiff (time difference).

An augmentation matrix for use with the variable timestep method can be generated by setting vtimestep=TRUE and augment=TRUE. Under default behavior, the augmentation matrix will include only the Tdiff combinations observed in the original vtimestep matrix, up to nreps. If you supply a vector Tdiff_fore, then the augmentation matrix will include or all possible combinations of the Tdiff values supplied in Tdiff_fore, even if they weren't in the original vtimestep matrix, up to nreps.

If generating forecasts, the output matrix will include a column for population if there is more than one, and will include a time column if time is supplied. The time increment is based on the minimum difference between timepoints. If generating forecasts using the variable timestep method, a vector of time units to forecast beyond the last timestep can be provided under Tdiff_fore.

Examples

yrand <- rnorm(20)
site <- rep(c("a","b"),each=10)
dfrand <- data.frame(firstvar=rnorm(20),secondvar=rnorm(20))
makelags(y=yrand,E=3,tau=1)
#>             var_1       var_2       var_3
#>  [1,]          NA          NA          NA
#>  [2,]  0.70215013          NA          NA
#>  [3,]  1.00538353  0.70215013          NA
#>  [4,] -0.50796991  1.00538353  0.70215013
#>  [5,] -0.03943715 -0.50796991  1.00538353
#>  [6,] -0.41887096 -0.03943715 -0.50796991
#>  [7,]  1.17678935 -0.41887096 -0.03943715
#>  [8,] -0.31493899  1.17678935 -0.41887096
#>  [9,] -0.16270667 -0.31493899  1.17678935
#> [10,] -0.14335065 -0.16270667 -0.31493899
#> [11,]  2.21078445 -0.14335065 -0.16270667
#> [12,] -0.14938012  2.21078445 -0.14335065
#> [13,] -0.99656848 -0.14938012  2.21078445
#> [14,] -1.02036964 -0.99656848 -0.14938012
#> [15,] -0.85448351 -1.02036964 -0.99656848
#> [16,] -0.36818894 -0.85448351 -1.02036964
#> [17,] -0.49876778 -0.36818894 -0.85448351
#> [18,] -0.35470238 -0.49876778 -0.36818894
#> [19,] -2.30287224 -0.35470238 -0.49876778
#> [20,] -0.08288528 -2.30287224 -0.35470238
makelags(y=dfrand,E=2,tau=2)
#>        firstvar_2  firstvar_4 secondvar_2 secondvar_4
#>  [1,]          NA          NA          NA          NA
#>  [2,]          NA          NA          NA          NA
#>  [3,] -1.28171716          NA  1.90576238          NA
#>  [4,] -0.79072358          NA -0.39850660          NA
#>  [5,] -1.13166862 -1.28171716 -1.07623775  1.90576238
#>  [6,]  0.24840636 -0.79072358 -1.42619268 -0.39850660
#>  [7,]  0.65202283 -1.13166862  1.27483659 -1.07623775
#>  [8,]  0.29643358  0.24840636 -0.72167029 -1.42619268
#>  [9,]  0.07790172  0.65202283  1.69810777  1.27483659
#> [10,] -1.22417268  0.29643358  0.08598211 -0.72167029
#> [11,]  0.81843090  0.07790172  0.55514410  1.69810777
#> [12,]  1.00346221 -1.22417268  0.87082602  0.08598211
#> [13,]  1.10607660  0.81843090 -0.62535176  0.55514410
#> [14,] -0.25071046  1.00346221 -1.36459801  0.87082602
#> [15,] -0.01646439  1.10607660  0.33172662 -0.62535176
#> [16,]  1.23760787 -0.25071046 -0.42204714 -1.36459801
#> [17,] -1.31227260 -0.01646439 -0.06279544  0.33172662
#> [18,]  1.25240725  1.23760787 -0.13528722 -0.42204714
#> [19,]  0.79438602 -1.31227260 -0.24997233 -0.06279544
#> [20,] -0.50169160  1.25240725  1.02978189 -0.13528722
makelags(y=dfrand,pop=site,E=2,tau=1)
#>        firstvar_1  firstvar_2 secondvar_1 secondvar_2
#>  [1,]          NA          NA          NA          NA
#>  [2,] -1.28171716          NA  1.90576238          NA
#>  [3,] -0.79072358 -1.28171716 -0.39850660  1.90576238
#>  [4,] -1.13166862 -0.79072358 -1.07623775 -0.39850660
#>  [5,]  0.24840636 -1.13166862 -1.42619268 -1.07623775
#>  [6,]  0.65202283  0.24840636  1.27483659 -1.42619268
#>  [7,]  0.29643358  0.65202283 -0.72167029  1.27483659
#>  [8,]  0.07790172  0.29643358  1.69810777 -0.72167029
#>  [9,] -1.22417268  0.07790172  0.08598211  1.69810777
#> [10,]  0.81843090 -1.22417268  0.55514410  0.08598211
#> [11,]          NA          NA          NA          NA
#> [12,]  1.10607660          NA -0.62535176          NA
#> [13,] -0.25071046  1.10607660 -1.36459801 -0.62535176
#> [14,] -0.01646439 -0.25071046  0.33172662 -1.36459801
#> [15,]  1.23760787 -0.01646439 -0.42204714  0.33172662
#> [16,] -1.31227260  1.23760787 -0.06279544 -0.42204714
#> [17,]  1.25240725 -1.31227260 -0.13528722 -0.06279544
#> [18,]  0.79438602  1.25240725 -0.24997233 -0.13528722
#> [19,] -0.50169160  0.79438602  1.02978189 -0.24997233
#> [20,] -0.04270665 -0.50169160  0.73691607  1.02978189
makelags(y=yrand,pop=site,E=2,tau=3,forecast = TRUE, yname="SomeName",time=c(1:10,1:10))
#>   time pop  SomeName_3 SomeName_6
#> 1   11   a -0.16270667 -0.4188710
#> 2   12   a -0.14335065  1.1767893
#> 3   13   a  2.21078445 -0.3149390
#> 4   11   b -2.30287224 -0.3681889
#> 5   12   b -0.08288528 -0.4987678
#> 6   13   b  1.42204107 -0.3547024
dfrand2 <- cbind.data.frame(Time=c(1:10,1:10),Site=site,dfrand)
makelags(data=dfrand2, y=c("firstvar","secondvar"),pop="Site",E=2,tau=3)
#>        firstvar_3  firstvar_6 secondvar_3 secondvar_6
#>  [1,]          NA          NA          NA          NA
#>  [2,]          NA          NA          NA          NA
#>  [3,]          NA          NA          NA          NA
#>  [4,] -1.28171716          NA  1.90576238          NA
#>  [5,] -0.79072358          NA -0.39850660          NA
#>  [6,] -1.13166862          NA -1.07623775          NA
#>  [7,]  0.24840636 -1.28171716 -1.42619268   1.9057624
#>  [8,]  0.65202283 -0.79072358  1.27483659  -0.3985066
#>  [9,]  0.29643358 -1.13166862 -0.72167029  -1.0762378
#> [10,]  0.07790172  0.24840636  1.69810777  -1.4261927
#> [11,]          NA          NA          NA          NA
#> [12,]          NA          NA          NA          NA
#> [13,]          NA          NA          NA          NA
#> [14,]  1.10607660          NA -0.62535176          NA
#> [15,] -0.25071046          NA -1.36459801          NA
#> [16,] -0.01646439          NA  0.33172662          NA
#> [17,]  1.23760787  1.10607660 -0.42204714  -0.6253518
#> [18,] -1.31227260 -0.25071046 -0.06279544  -1.3645980
#> [19,]  1.25240725 -0.01646439 -0.13528722   0.3317266
#> [20,]  0.79438602  1.23760787 -0.24997233  -0.4220471
makelags(data=dfrand2, y=c("firstvar","secondvar"),pop="Site",E=2,tau=3,
forecast = TRUE,time="Time")
#>   Time Site  firstvar_3  firstvar_6 secondvar_3 secondvar_6
#> 1   11    a -1.22417268  0.65202283  0.08598211  1.27483659
#> 2   12    a  0.81843090  0.29643358  0.55514410 -0.72167029
#> 3   13    a  1.00346221  0.07790172  0.87082602  1.69810777
#> 4   11    b -0.50169160 -1.31227260  1.02978189 -0.06279544
#> 5   12    b -0.04270665  1.25240725  0.73691607 -0.13528722
#> 6   13    b  1.80350446  0.79438602  0.92515979 -0.24997233